--- lib/features/survey/service/data.dart
@@ -1,128 +1,0 @@
-/// Survey data service.
-//
-// Time-stamp: <Thursday 2024-12-19 13:33:06 +1100 Graham Williams>
-//
-/// Copyright (C) 2025, Software Innovation Institute, ANU
-///
-/// Licensed under the GNU General Public License, Version 3 (the "License");
-///
-/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
-//
-// This program is free software: you can redistribute it and/or modify it under
-// the terms of the GNU General Public License as published by the Free Software
-// Foundation, either version 3 of the License, or (at your option) any later
-// version.
-//
-// This program is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
-// details.
-//
-// You should have received a copy of the GNU General Public License along with
-// this program.  If not, see <https://www.gnu.org/licenses/>.
-///
-/// Authors: Ashley Tang
-
-library;
-
-import 'dart:convert';
-
-import 'package:flutter/material.dart';
-
-import 'package:solidpod/solidpod.dart';
-
-/// Survey data service.
-///
-/// This service handles the retrieval of survey data from remote POD storage.
-/// Ensures all data is fetched, sorted and ready for use.
-
-class SurveyDataService {
-  // Fetch from directory where blood pressure-related survey data resides.
-
-  static const String bpDir = 'healthpod/data/bp';
-
-  /// Fetches survey data from POD, ensuring it is sorted by timestamp.
-  ///
-  /// Can potentially fetch from local storage as well, but this is omitted for now,
-  /// as we assume all relevant bp data is stored in POD or uploaded from local already.
-  /// Acts as main entry point.
-
-  static Future<List<Map<String, dynamic>>> fetchAllSurveyData(
-      BuildContext context) async {
-    List<Map<String, dynamic>> allData = [];
-
-    // Fetch POD data.
-
-    if (context.mounted) {
-      final podData = await fetchPodSurveyData(context);
-      allData.addAll(podData);
-    }
-
-    // Sort all data by timestamp.
-
-    allData.sort((a, b) => DateTime.parse(a['timestamp'])
-        .compareTo(DateTime.parse(b['timestamp'])));
-
-    return allData;
-  }
-
-  /// Fetches survey data from POD storage.
-
-  static Future<List<Map<String, dynamic>>> fetchPodSurveyData(
-      BuildContext context) async {
-    List<Map<String, dynamic>> podData = [];
-    try {
-      // Get the directory URL for the bp folder.
-
-      final dirUrl = await getDirUrl(bpDir);
-
-      // Get resources in the container.
-
-      final resources = await getResourcesInContainer(dirUrl);
-
-      debugPrint('SubDirs: |${resources.subDirs.join('|')}|');
-      debugPrint('Files  : |${resources.files.join('|')}|');
-
-      // Process each file in the directory.
-
-      for (var fileName in resources.files) {
-        if (!fileName.endsWith('.enc.ttl')) continue;
-
-        // Construct the full path including healthpod/data/bp.
-
-        final filePath = '$bpDir/$fileName';
-
-        if (!context.mounted) break;
-
-        // Read the file content.
-
-        final result = await readPod(
-          filePath,
-          context,
-          const Text('Reading survey data'),
-        );
-
-        // Handle the response based on its type.
-
-        if (result != SolidFunctionCallStatus.fail &&
-            result != SolidFunctionCallStatus.notLoggedIn) {
-          try {
-            // The result is the JSON string directly.
-
-            final data = json.decode(result.toString());
-            podData.add(data);
-          } catch (e) {
-            debugPrint('Error parsing file $fileName: $e');
-            debugPrint('Content: $result');
-          }
-        } else {
-          debugPrint('Failed to read file $fileName: $result');
-        }
-      }
-    } catch (e) {
-      debugPrint('Error fetching POD survey data: $e');
-      debugPrint('Error details: ${e.toString()}');
-    }
-    return podData;
-  }
-}
